requirements.txt
Flask==3.0.0
pdfminer.six==20221105
google-generativeai==0.2.2
gunicorn==21.2.0